Update README.md
The purpose of this project is to highlight key finding about Silo's coffee shops and help the CEO make informed decision that will help the company increase revenue and improve product performance.
Using the data and market insights to develop new strategies. Miro was used to brainstorm and build a roadmap for the project. SQL was used to draw data and excel was used to analysis the data. The project is presented on PowerPoint.